Fix disable killing process of MacOS for Github Action

+++ system/macos.sh
@@ -1458,11 +1458,14 @@ ExtraSettings
 ## ----------------------------------------
 ##	Cache Clear
 ## ----------------------------------------
-for app in \
-	"cfprefsd" \
-	"Activity Monitor" "Address Book" "Calendar" \
-	"Contacts" "Dock" "Finder" "Mail" "Messages" \
-	"SystemUIServer" "Terminal" "Transmission" "iCal"
-do
-	killall ${app}
-done
+TESTMODE=$1
+if ! ${TESTMODE}; then
+	for app in \
+		"cfprefsd" \
+		"Activity Monitor" "Address Book" "Calendar" \
+		"Contacts" "Dock" "Finder" "Mail" "Messages" \
+		"SystemUIServer" "Terminal" "Transmission" "iCal"
+	do
+		killall ${app}
+	done
+fi
